refactor: extract history reconstruction service and static helpers from ChatViewModel (#21841)
